requirements.txt
streamlit==1.46.1
python-dotenv==1.1.1
requests==2.32.4
qrcode==8.2
pillow==11.3.0
pytz==2025.2
letter==0.5
reportlab==4.4.2